Smarter unexpected sequence report
* Rework unexpected sequence writing

Rather than sample unexpected sequences by % of each column barcode's
unexpected sequences, we set a maximum number of row barcodes for which
we are willing to track unexpected sequences. When reading the
unexpected sequence cache, we take 1 barcode from each shard in turn,
until we have read all of them or established a sample of that size. All
further unexpected sequences that are not found in that set are not
tracked.

This lets us bound the memory that PoolQ uses for its unexpected
sequence reporting, without needing to individually track the sizes of
individual shards. It should also perform better than the previous
strategy since it reads the cache only once.

* Do not track shard sizes

* Set version to 3.10.0-SNAPSHOT

* Update changelog, manual, and readme

* Fix makefile

* Use `.size`

* Simpler breadth-first iterator